refactor: move KeySet to common-procedure
